Move unit test that depends on Redis to the integration test suite
